refactor(cli): deduplicate makeAbsolute
