+++ src/cli/utl/makeAbsolute.js
@@ -1,0 +1,11 @@
+const path = require('path');
+
+module.exports = (pFilename) => {
+    let lRetval = pFilename;
+
+    if (!path.isAbsolute(pFilename)) {
+        lRetval = path.join(process.cwd(), pFilename);
+    }
+    return lRetval;
+
+};
