+++ src/cli/utl/validateFileExistence.js
