+++ test/cli/utl/makeAbsolute.spec.js
@@ -1,0 +1,17 @@
+const path         = require('path');
+const expect       = require('chai').expect;
+const makeAbsolute = require('../../../src/cli/utl/makeAbsolute');
+
+describe("makeAbsolute", () => {
+    it("leaves absolute path names alone", () => {
+        expect(
+            makeAbsolute("/hallo/dit/is/een/absoluut/pad")
+        ).to.equal("/hallo/dit/is/een/absoluut/pad");
+    });
+
+    it("puts the current working directory in front of non-absolute paths", () => {
+        expect(
+            makeAbsolute("pad")
+        ).to.equal(path.join(process.cwd(), "pad"));
+    });
+});
